Add: WAIT/RETAIN dependency attribute split (a2a3 tmr, issue #1375)
One physical dependency edge coupled WAIT (readiness) and RETAIN
(lifetime); split them into orthogonal 2-bit flags so later transitive
reduction can safely downgrade WAIT|RETAIN -> RETAIN-only. This commit
ships the attribute split only (a2a3 tensormap_and_ringbuffer) — no
deduplication. host_build_graph is untouched (upstream removed its
fanin spill/for_each machinery); a5 is deferred.

- PTO2TaggedSlotPtr wraps slot pointers tagged with PTO2DepFlags in
  the low bits (fanin inline array, spill entries, fanout dep-list
  entries): same sizes/layouts, trivially default constructible, and
  misuse (dereferencing a still-tagged pointer) is a compile error.
- Provenance mapping: creator (owner_task_id) = WAIT|RETAIN, tensormap
  modifier = WAIT, explicit = WAIT|RETAIN; duplicate discoveries
  OR-merge via an epoch|index seen array instead of first-wins.
- Pin protocol: WAIT-only pins release in the producer completion
  fanout walk (scheduler thread, batched CONSUMED check), at wiring
  for pre-completed producers, or in the all-completed fast path;
  on_task_release releases RETAIN edges only; early-dispatch candidacy
  target is fanin_count - 1 (WAIT edges), not fanin_actual_count.
- Retain-first fanin ordering (inline fanins) with
  fanin_retain_count in the pre-predicate payload slack (offset 564,
  layout otherwise unchanged); spilled fanins keep discovery order as
  a deliberate full-walk fallback. for_each_fanin_storage internalizes
  traversal via a PTO2DepFlags select with two compile-time callback
  forms (unfiltered with flags, or flagless exact-kind match).
- Alignment guards: static_assert(alignof >= 4) plus cold-path
  always_assert on the suballocated slot_states base.
- dep_gen replay: emit gains the flags argument; oracle ignores it,
  annot mirror unchanged.
- docs/two-kinds-of-dep.md: design, the measured-and-removed depth-1
  reduction prototype (with cost decomposition for the follow-up
  work), and the open batch_paged_attention orch delta (~+2-3%).

Verified: sim a2a3 tmr suite 47/47; onboard (task-submit) 9/9 core
cases; benchmark 8/8, only batch_paged_attention shows a reproducible
+2~3% orch delta, all other cases within noise.

+++ src/a2a3/runtime/tensormap_and_ringbuffer/runtime/pto_orchestrator.h
@@ -69,7 +69,10 @@ struct PTO2OrchestratorState {
 
     // === PER-RING RESOURCES ===
     PTO2RingSet rings[PTO2_MAX_RING_DEPTH];
-    uint32_t *fanin_seen_epoch[PTO2_MAX_RING_DEPTH];
+    // Per-slot dedup record for fanin discovery: high 32 bits = seen epoch,
+    // low 32 bits = builder edge index of the first claim (valid iff the
+    // epoch matches), so duplicate discoveries OR-merge flags in place.
+    uint64_t *fanin_seen_epoch[PTO2_MAX_RING_DEPTH];
     uint32_t fanin_seen_current_epoch{1};
 
     // === TENSOR MAP (Private) ===
